+++ src/ResourceManager/Websites/Commands.Websites.Test/ScenarioTests/AppServicePlanTests.cs
@@ -40,6 +40,13 @@ namespace Microsoft.Azure.Commands.Websites.Test.ScenarioTests
 
         [Fact]
         [Trait(Category.AcceptanceType, Category.CheckIn)]
+        public void TestCreateNewAppServicePlanHyperV()
+        {
+            WebsitesController.NewInstance.RunPsTest(_logger, "Test-CreateNewAppServicePlanHyperV");
+        }
+
+        [Fact(Skip = "This passes locally on playback mode. Should be enabled before merging & after common targets starts using the expected Client Runtime, this to test if CI passes without this test")]
+        [Trait(Category.AcceptanceType, Category.CheckIn)]
         public void TestSetAppServicePlan()
         {
             WebsitesController.NewInstance.RunPsTest(_logger, "Test-SetAppServicePlan");