+++ src/ResourceManager/Websites/Commands.Websites/ChangeLog.md
@@ -18,7 +18,10 @@
         - Additional information about change #1
 -->
 ## Current Release
-* Updating to use the latest .NET SDK version (2.0.0)
+* Updating to use the latest .NET SDK version (2.0.0) https://www.nuget.org/packages/Microsoft.Azure.Management.Websites
+* New-AzureRmAppServicePlan -HyperV switch is added for create app service plan with windows container
+* New-AzureRmWebApp/ New-AzureRmWebAppSlot/ Set-AzureRmWebApp/ Set-AzureRmWebAppSlot - New parameters (–ContainerRegistryUser “<string>” -ContainerRegistryPassword “<secureString>” -EnableContainerContinuousDeployment) added for creating and managing windows container app
+
 
 ## Version 5.0.9
 * Fixed issue with default resource groups not being set.